HSEARCH-3269 Move component tests to IT test module
