+++ batch-jsr352/core/src/main/java/org/hibernate/search/batch/jsr352/core/massindexing/step/lucene/spi/EntityReader.java
@@ -149,10 +149,7 @@ public class EntityReader extends AbstractItemReader {
 	public EntityReader() {
 	}
 
-	/**
-	 * Constructor for unit test
-	 */
-	EntityReader(String serializedCacheMode,
+	public EntityReader(String serializedCacheMode,
 			String entityName,
 			String serializedEntityFetchSize,
 			String serializedCheckpointInterval,
